+++ instrumentation/opensearch/opensearch-java-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/opensearch/v3_0/OpenSearchTransportInstrumentation.java
@@ -72,7 +72,10 @@ class OpenSearchTransportInstrumentation implements TypeInstrumentation {
 
     @Nullable
     public static AdviceScope start(
-        Object request, Endpoint<Object, Object, Object> endpoint, JsonpMapper jsonpMapper) {
+        OpenSearchTransport transport,
+        Object request,
+        Endpoint<Object, Object, Object> endpoint,
+        JsonpMapper jsonpMapper) {
       Context parentContext = Context.current();
 
       String queryBody = null;
@@ -84,7 +87,11 @@ class OpenSearchTransportInstrumentation implements TypeInstrumentation {
 
       OpenSearchRequest otelRequest =
           OpenSearchRequest.create(
-              endpoint.method(request), endpoint.requestUrl(request), queryBody);
+              endpoint.method(request),
+              endpoint.requestUrl(request),
+              queryBody,
+              OpenSearchServerTargets.address(transport),
+              OpenSearchServerTargets.port(transport));
 
       if (!instrumenter().shouldStart(parentContext, otelRequest)) {
         return null;
@@ -138,7 +145,8 @@ class OpenSearchTransportInstrumentation implements TypeInstrumentation {
         @Advice.This OpenSearchTransport openSearchTransport,
         @Advice.Argument(0) Object request,
         @Advice.Argument(1) Endpoint<Object, Object, Object> endpoint) {
-      return AdviceScope.start(request, endpoint, openSearchTransport.jsonpMapper());
+      return AdviceScope.start(
+          openSearchTransport, request, endpoint, openSearchTransport.jsonpMapper());
     }
 
     @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class, inline = false)
@@ -160,7 +168,8 @@ class OpenSearchTransportInstrumentation implements TypeInstrumentation {
         @Advice.This OpenSearchTransport openSearchTransport,
         @Advice.Argument(0) Object request,
         @Advice.Argument(1) Endpoint<Object, Object, Object> endpoint) {
-      return AdviceScope.start(request, endpoint, openSearchTransport.jsonpMapper());
+      return AdviceScope.start(
+          openSearchTransport, request, endpoint, openSearchTransport.jsonpMapper());
     }
 
     @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class, inline = false)